requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.1
django-environ==0.11.2
Faker==22.6.0
gunicorn==21.2.0
humanize==4.9.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
